Make Pose3d use JNI
